Remove coverage file and have everything inside the fixtures.ts file.
